[Java] Fix truncation of the linger timeout in the ChannelUriStringBuilder. (#1208)
* [Java] Fix truncation of the linger timeout in the ChannelUriStringBuilder.

* [Java] Add more tests for linger parameter.

+++ aeron-client/src/main/java/io/aeron/ChannelUriStringBuilder.java
@@ -1971,7 +1971,7 @@ public final class ChannelUriStringBuilder
 
         if (null != linger)
         {
-            sb.append(LINGER_PARAM_NAME).append('=').append(linger.intValue()).append('|');
+            sb.append(LINGER_PARAM_NAME).append('=').append(linger).append('|');
         }
 
         if (null != alias)
